refactor: Art를 Arts 폴더로 이동
