--- src/main/kotlin/octoping/ticketing/api/controller/art/request/ArtListResponseDTO.kt
@@ -1,9 +1,0 @@
-package octoping.ticketing.api.controller.art.request
-
-import octoping.ticketing.api.controller.art.schema.ArtItemSchema
-
-data class ArtListResponseDTO(
-    val arts: List<ArtItemSchema>,
-    val page: Int,
-) {
-}
