+++ src/main/kotlin/octoping/ticketing/api/controller/arts/request/ArtInfoResponseDTO.kt
@@ -1,4 +1,4 @@
-package octoping.ticketing.api.controller.art.request
+package octoping.ticketing.api.controller.arts.request
 
 import java.time.LocalDate
